removed debug, added header

+++ src/api/actions/DropsUserUpdater.class.php
@@ -29,7 +29,7 @@ class DropsUserUpdater extends DropsUserAction
             'lastName' => get_user_meta($userId, 'last_name', true),
             'mobilePhone' => get_user_meta($userId, 'mobile', true),
             'placeOfResidence' => get_user_meta($userId, 'residence', true),
-            'birthday' => (float)get_user_meta($userId, 'birthday', true),
+            'birthday' => (int)get_user_meta($userId, 'birthday', true),
             'sex' => get_user_meta($userId, 'gender', true),
         );
